2861: Add PrivacyCheckbox as own component separate from Checkbox component in native
